ARM: OMAP2+: AM33XX: toggle to exit self refresh in resume from from deep sleep
When the device is in DS0, all EMIF context is lost. When the device
wakes and the EMIF clock is reenabled, the EMIF registers will be at
their default state. So we must get the EMIF back into the state it
was in during self-refresh.

When we wakeup, we restore POWER_MGMT_CTRL back to a lp_mode=2 (which
is self-refresh), then you write a lp_mode=0 to get it out of self
refresh and finally restore the original setting with which we enter
into suspend path.

As part of the change - dont use r0 as scratchpad for delay loop since
it stores the base address for emif

[nm@ti.com: minor edits and changelog]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Russ Dill <Russ.Dill@ti.com>

(cherry picked from commit a9ae92a0ae384a78648bd271342ee2c4b2e37376)
(cherry picked from commit 6d7a7c092e63e2920367c144f8e8cb8bfad5bf9b)