bus: omap_l3_noc: Add support for discountinous flag mux input numbers
On DRA7, unlike on OMAP4 and OMAP5, the flag mux input numbers used to indicate
the source of errors are not continous. Have a way in the driver to catch these
and WARN the user of the flag mux input thats either undocumented or wrong.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>
(cherry picked from commit 64832649472b782a2b37dc5f4bb984386290a767)