bus: omap_l3_noc: support variable mem resource
There are SoC's (like AM43x) that can have less than 3 clockdomains.
Modify the driver so that it can handle different number of
clockdomains to prepare for AM43x support.

While at it use managed device resources and avoid a kfree in error path
which should not be done as the memory is not an allocated one.

Signed-off-by: Afzal Mohammed <afzal@ti.com>
(cherry picked from commit 7bc775f25a6819f21810bea40792f4af367688d3)