ARM: AM33XX: Remove unnecessary delays from suspend/resume path
No delays are mentioned as being required in any of the TRM documentation.
Instead, wait for the EMIF to indicate that it is ready.

This is needed for am33xx CPU idle support, which is sensitive to return
to active status.

Signed-off-by: Russ Dill <russ.dill@ti.com>
(cherry picked from commit dcad654f4ccd5466934e2359744d9f4a0ce57216)