ARM: config: Add HAVE_ARM_SCU for AM43XX
CONFIG_HAVE_ARM_SCU only gets selected if CONFIG_SMP is selected in an OMAP
system however AM43XX needs this option regardless of CONFIG_SMP as it is
important for controlling power in the SoC. Without this suspend will
not work as scu_power_mode becomes an empty function.

AM43XX requires SCU CPU Power Status bits for A9 core to be set to off mode
in order for the MSTANDBY signal to be asserted from the MPU during WFI.
This signal is used by the PRCM to determine when it is appropriate to
clock gate the MPU clock domain and also trigger M3_IRQ2 to tell the
wkup_m3 firmware to put the part into the desired sleep more.