overlayfs: dont take mnt_want_write for lifetime of the overlay
The lockdep error is due to lazyness on my part and trying to hold
mnt_want_write() on upper for the lifetime of the overlay.  That doesn't
work well with the freeze protection part (sb_start_write()) because then
upper can't be frozen while the overlay is mounted, which is not what we
want, I think.

Reported-by: David Howells <dhowells@redhat.com>
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>