spi: qup: Ensure done detection
This patch fixes an issue where a SPI transaction has completed, but the done
condition is missed.  This occurs because at the time of interrupt the
MAX_INPUT_DONE_FLAG is not asserted.  However, in the process of reading blocks
of data from the FIFO, the last portion of data comes in.

The opflags read at the beginning of the irq handler no longer matches the
current opflag state.  To get around this condition, the block read function
should update the opflags so that done detection is correct after the return.

Signed-off-by: Andy Gross <agross@codeaurora.org>
(cherry picked from commit 4faba89e3ffbb1c5f6232651375b9b3212b50f02)

Change-Id: I415fab265a4451a4d614010a8c0866956831530a
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>