watchdog: qcom: set WDT_BARK_TIME register offset to one second less of bite time
Currently WDT_BARK_TIME register offset is not configured with bark
timeout during wdt_start,and it is taking bark timeout's default value.
For some versions of TZ (secure mode) will consider a BARK the same
as BITE and reset the board.

So instead let's just configure the BARK time to be less than a second
of the bite timeout so the board does not reset in this scenario

Change-Id: Ie09850ad7e0470ed721e6924911ca2a81fd9ff8a
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>