spi: qup: properly detect extra interrupts
It's possible for a SPI transaction to complete and get another
interrupt and have it processed on the same spi_transfer before the
transfer_one can set it to NULL.

This masks unexpected interrupts, so let's set the spi_transfer to
NULL in the interrupt once the transaction is done. So we can
properly detect these bad interrupts and print warning messages.

Change-Id: I0e70ed59fb50e5c48a72a38f74bd178b17c9f24d
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>