ipq806x: qcom: Rename model from AP148-NAND to AP148 alone
The device-tree model will be used to form the board name,
existing model name AP148-NAND will be treated as a separate
board which is creating the board name as unknown, hence the
default uci configs are not getting generated.

In order to fix this, instead of adding a separate board support,
we can change the device-tree model name as AP148.

Change-Id: Ifa2f96a4bd7b39e0c1fd166a49d99f66f8ae14cf
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>