ipq40xx: I2C: Block mode Fixes
I2C block mode related fixes including those to make it work
for lengths greater than 256 bytes.

I2C can transfer only lengths of 256 bytes per block due to
the len parameter in the tag which is 8 bit wide. Hence a
given datalen is split into multiple blocks each of len
256 bytes. For block mode to work, it requires the total len
per block inclusive of the tags to be a multiple of the HW
Block size.

For read transactions, block mode requires that the tag write
is aligned to word boundary.

Change-Id: I3cf8fbfaf569c77262e13b3a041684e9baf9cd9d
Signed-off-by: Charanya Venkatraman <charanya@codeaurora.org>