staging: wilc1000: Make use of APIs in ieee80211.h
This patch series contains changes to address the review feedback
comments [1]. It mainly contains the changes to have clear names for
the constants. Most of the constants names are taken care, few are left
unchanged in hope with another cleanup those macros might be deleted. After
further cleanup will revisit if anything else is required to address this.

[1]. - https://www.spinics.net/lists/linux-wireless/msg178706.html
     - https://www.spinics.net/lists/linux-wireless/msg178704.html

Signed-off-by: Adham Abozaeid <adham.abozaeid@microchip.com>