staging: wilc1000: Send reset request to WILC CPU instead of resetting it
Resetting WILC's CPU ungracefuly can leave it in an unknown state
specially if it was in sleep mode, and can cause the part to fail
permanently on the long run.
The drive should instead send a sleep request for the FW to deinit.
Requires a matching change in the FW.

Signed-off-by: Adham Abozaeid <adham.abozaeid@microchip.com>