usb: musb: musb_dsps: Fix host mode suspend
On AM335x, VBUS is not maintained during suspend when musb is in host mode.
On resume, all the connected devices will be disconnected and then will
be enumerated again.
This happens because MUSB_DEVCTL is cleared during suspend.

Use the quirk MUSB_PRESERVE_SESSION to preseve MUSB_DEVCTL during suspend.

Signed-off-by: Takeyoshi Kikuchi <kikuchi@centurysys.co.jp>