i2c: at91: Workaround to fix GPIO pinmux
After the SCL and SDA GPIOs are received, the pins are muxed as GPIOs,
without informing the pinctrl subsystem. This breaks the I2C bus, since
the I2C is no longer using those pins. To mux the pins back to I2C, we
tell pinctrl to change the state to gpio first and then to default
(I2C).

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>