net: macb: free resources in case of failure
In case macb_open() fails and user space tries to "ifup" the again
the interface after this failure, the napi_enable() will reach a
BUG_ON() instruction because the napi wasn't disabled the the
previous macb_open() failure. So, add failure path in macb_open()
that reverts the previously done operations.

Fixes: b2b041417299 ("net: macb: convert to phylink")
Tested-by: Cristian Birsan <cristian.birsan@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>