wilc1000: use wait_for_completion_interruptible_timeout for completion
The default driver implementation blocks the TXQ task until the upper
layer posts Tx packet. If the interface is idle, the thread will be
in waiting state forever which makes the kernel to think that the
task is hung.

Use wait_for_completion_interruptible_timeout() inplace of
wait_for_completion_timeout() to solve this issue.

Signed-off-by: Ajay Singh <ajay.kathat@microchip.com>
Signed-off-by: Sripad Balwadgi <Sripad.Balwadgi@microchip.com>